Refactor database configuration to use MySQL instead of SQLite
