+++ .env
@@ -27,12 +27,12 @@ LOG_LEVEL=debug
 # DB_USERNAME=root
 # DB_PASSWORD=kJTKJFXjPwoIQNahYbsznTWaTvwMlfSG
 
-DB_CONNECTION=sqlite
-# DB_HOST=mysql
-# DB_PORT=3306
-# DB_DATABASE=laravel
-# DB_USERNAME=sail
-# DB_PASSWORD=password
+DB_CONNECTION=mysql
+DB_HOST=mysql
+DB_PORT=3306
+DB_DATABASE=laravel
+DB_USERNAME=sail
+DB_PASSWORD=password
 
 SESSION_DRIVER=database
 SESSION_LIFETIME=120
