+++ app/Models/ProjectPayItemModel.php
@@ -5,7 +5,7 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
-class ContractPayItemModel extends Model
+class ProjectPayItemModel extends Model
 {
     use HasFactory;
