+++ resources/js/Pages/Contract/ContractPage.jsx
@@ -12,8 +12,8 @@ import {
 import { ChevronDown } from "lucide-react";
 import { useState } from "react";
 import _ from "lodash";
-import PastContractsTable from "@/Componentss/contract/PastContractsTable";
-import ContractCard from "@/Componentss/contract/ContractCard";
+import PastContractsTable from "@/Components/contract/PastContractsTable";
+import ContractCard from "@/Components/contract/ContractCard";
 import { Input } from "@/Components/ui/input";
 import { router } from "@inertiajs/react";
