+++ resources/js/Pages/Welcome.jsx
@@ -1,5 +1,5 @@
 
-import ApplicationLogo from "@/Componentss/ApplicationLogo";
+import ApplicationLogo from "@/Components/ApplicationLogo";
 import { Button } from "@/Components/ui/button";
 import {
     Card,
